deplacement de la logique de code pour le module des offres vers un controller pour les offres
