+++ LivrETS/Views/Shared/_header.cshtml
@@ -76,7 +76,7 @@
                     <li>@Html.ActionLink("Accueil", "Index", "Home")</li>
                     @if (Request.IsAuthenticated)
                     {
-                        <li>@Html.ActionLink("Vendre", "Sell", "Home")</li>
+                        <li>@Html.ActionLink("Vendre", "Create", "Offer")</li>
 
                         if (User.IsInRole("Administrator") || User.IsInRole("Clerk"))
                         {
